components/odrive/requirements.txt
paho-mqtt==1.6.1
odrive==0.6.3.post0
pyyaml==6.0
